pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vacuumDecay"
version = "0.1.0"
dependencies = [
    "torch",
    "numpy",
    "flask",
    "flask-socketio",
    "networkx",
    "pillow"
]

[tool.setuptools]
packages = ["vacuumDecay"]

